Makefile: create parent directories
cp requires the destination directories to already exist. If they do not
exist, it will fail. When packaging applications, it's common they are
installed in an empty directory where the expected directory structure
does not exist yet.

Use `install -D` to copy the files to copy the files so that parent
diretories are automatically created.

Signed-off-by: Kevin Daudt <me@ikke.info>