filter-repo: notify users when we remove the origin remote
Folks often run commands without reading the docs, but they are more
likely to read the output of the command.  Notify users that we remove
the origin remote, and point to a new section in the docs that explains
why we do so.  That should make it easier for them to discover the
potential pitfalls that await, as well as help them more quickly
discover how to proceed.

Signed-off-by: Elijah Newren <newren@gmail.com>