filter-repo: add a new get_parent_hashes() convenience function
We want to add the ability to show the first changed commit(s) in a
rewrite, i.e. those commits that were rewritten but which did not have
any parents that were rewritten.  And we want to be able to update these
on subsequent runs of filter-repo when users rewrite their repository in
multiple steps.

To facilitate this, add a get_parent_hashes() function to the
AncestryGraph class.

Signed-off-by: Elijah Newren <newren@gmail.com>