filter-repo: add tracking and recording of first changed commit(s)
Compute the set of first changed commits in a rewrite, that is, those
commits that were rewritten which had no parents that were rewritten.
There will typically be just one of these in a rewrite, making it a
good candidate for a "ban commit" hook to prevent the old history from
being repushed to a repository.  Store these in
$GIT_DIR/filter-repo/first-changed-commits.

Signed-off-by: Elijah Newren <newren@gmail.com>