filter-repo: fix corner case involving bad pruning of evil degenerate merge
When a former merge commit either has its entire first parent history
removed, or a former merge commit has enough of its first parent history
removed that its new first parent is an ancestory of its second parent,
we consider whether the merge itself should be pruned.

The basic logic is a merge can be removed in either of these cases if
the merge introduces no file changes relative to its remaining parent.
However, our logic for determining if the merge introduced no file
changes relative to its remaining parent was faulty in the case of some
merges that had changes beyond what a simple merge would imply (i.e.
merges that had semantic conflict fixes or "evil" changes snuck in).
To properly address those, we need to instead work with the list of
changes relative to the new proposed parent.  Add code to handle this,
and mark the relevant testcase as passing.

Signed-off-by: Elijah Newren <newren@gmail.com>